+++ config/bootstrap.php
@@ -117,8 +117,36 @@ Log::config('users', [
 ]);
 
 /**
- * Adds `isAdmin()` detector for the request
+ * Adds `isAdmin()` detector
  */
 Request::addDetector('admin', function ($request) {
     return $request->param('prefix') === 'admin';
+});
+
+/**
+ * Adds `isBanned()` detector.
+ * It checks if the user's IP address is banned.
+ */
+Request::addDetector('banned', function ($request) {
+    $banned = config('security.banned_ip');
+
+    /**
+     * The IP address is allowed if:
+     *  - the list of banned ip is empty;
+     *  - is localhost;
+     *  - the IP address has already been verified.
+     */
+    if(!$banned || is_localhost() || $request->session()->read('allowed_ip'))
+        return FALSE;
+    
+	//Replaces asteriskes
+    $banned = preg_replace('/\\\\\*/', '[0-9]{1,3}', array_map('preg_quote', is_array($banned) ? $banned : [$banned]));
+
+    if(preg_match(sprintf('/^(%s)$/', implode('|', $banned)), $request->clientIp())) {
+        return TRUE;
+    }
+		
+    //In any other case, saves the result in the session
+    $request->session()->write('allowed_ip', TRUE);
+    return FALSE;
 });