+++ src/Controller/Admin/PhotosController.php
@@ -52,9 +52,6 @@ class PhotosController extends AppController {
             
             $this->set(compact('albums'));
 		}
-		
-		//See http://book.cakephp.org/2.0/en/core-libraries/components/security-component.html#disabling-csrf-and-post-data-validation-for-specific-actions
-		$this->Security->config('unlockedActions', 'upload');
 	}
 	
 	/**