+++ src/Controller/PostsController.php
@@ -153,7 +153,7 @@ class PostsController extends AppController {
 	
 	/**
 	 * Search posts
-	 * @uses MeCms\Controller\Component\SecurityComponent::checkLastSearch()
+	 * @uses MeCms\Controller\AppController::_checkLastSearch()
 	 * @uses MeCms\Model\Table\PostsTable::checkIfCacheIsValid()
 	 */
 	public function search() {
@@ -162,7 +162,7 @@ class PostsController extends AppController {
 		if($pattern) {
 			//Checks if the pattern is at least 4 characters long
 			if(strlen($pattern) >= 4) {
-				if($this->Security->checkLastSearch($pattern)) {
+				if($this->_checkLastSearch($pattern)) {
 					$this->paginate['limit'] = config('frontend.records_for_searches');
 					
 					//Checks if the cache is valid