+++ src/Controller/UsersController.php
@@ -48,9 +48,6 @@ class UsersController extends AppController {
 		//Checks if the user is already logged in
 		if(!$this->request->isAction('logout') && $this->Auth->isLogged())
 			return $this->redirect(['_name' => 'dashboard']);
-		
-		//See http://book.cakephp.org/2.0/en/core-libraries/components/security-component.html#disabling-csrf-and-post-data-validation-for-specific-actions
-		$this->Security->config('unlockedActions', ['forgot_password', 'resend_activation', 'signup']);
 	}
 	
 	/**